+++ lib/rdf/datatype_ns.ex
@@ -1,9 +1,9 @@
 defmodule RDF.Datatype.NS do
-  @moduledoc false
-
-  # Since the capability of RDF.Vocabulary.Namespaces requires the compilation
-  # of the RDF.NTriples.Decoder and the RDF.NTriples.Decoder depends on RDF.Literals,
-  # we can't define the XSD namespace in RDF.NS.
+  @moduledoc !"""
+  Since the capability of RDF.Vocabulary.Namespaces requires the compilation
+  of the RDF.NTriples.Decoder and the RDF.NTriples.Decoder depends on RDF.Literals,
+  we can't define the XSD namespace in RDF.NS.
+  """
 
   use RDF.Vocabulary.Namespace