+++ lib/rdf/datatypes/double.ex
@@ -82,4 +82,7 @@ defmodule RDF.Double do
     "#{i}.#{f}E#{e}"
   end
 
+
+  def equal_value?(left, right), do: RDF.Numeric.equal_value?(left, right)
+
 end